Added client option modules to differenciate between nonAzure and Azure
